+++ src/Paket.Core/NugetConvert.fs
@@ -214,7 +214,7 @@ let createPackageRequirement (packageName, version, restrictions) dependenciesFi
                 VersionRequirement(VersionRange.Minimum <| SemVer.Parse "0", PreReleaseStatus.No)
             else
                 VersionRequirement(VersionRange.Exactly version, PreReleaseStatus.No)
-       ResolverStrategy = ResolverStrategy.Max
+       ResolverStrategy = Some ResolverStrategy.Max
        Settings = { InstallSettings.Default with FrameworkRestrictions = restrictions }
        Parent = PackageRequirementSource.DependenciesFile dependenciesFileName }
 
@@ -297,6 +297,7 @@ let createDependenciesFileR (rootDirectory : DirectoryInfo) nugetEnv mode =
             Paket.DependenciesFile(DependenciesFileParser.parseDependenciesFile dependenciesFileName newLines))
 
     if File.Exists dependenciesFileName then read() else create()
+    |> lift (fun d -> d.SimplifyFrameworkRestrictions())
 
 let convertPackagesConfigToReferences projectFileName packagesConfig =
     let referencesFile = ProjectFile.FindOrCreateReferencesFile(FileInfo projectFileName)